Fix: deployment of server depends on dev packages as testUtils includes them
By moving the credentialFile routines to a separate package, there is no longer a
dependency on a large number of dev packages a runtime
