+++ packages/server/src/models/items/storage/StorageDriverS3.test.ts
@@ -2,7 +2,8 @@
 // defined in the below config file. If the credentials are missing, all the
 // tests are skipped.
 
-import { afterAllTests, beforeAllDb, beforeEachDb, readCredentialFileSync } from '../../../utils/testing/testUtils';
+import { afterAllTests, beforeAllDb, beforeEachDb } from '../../../utils/testing/testUtils';
+import { readCredentialFileSync } from '../../../utils/testing/credentialUtils';
 import { StorageDriverConfig, StorageDriverMode, StorageDriverType } from '../../../utils/types';
 import { shouldDeleteContent, shouldNotCreateItemIfContentNotSaved, shouldNotUpdateItemIfContentNotSaved, shouldSupportFallbackDriver, shouldSupportFallbackDriverInReadWriteMode, shouldThrowNotFoundIfNotExist, shouldUpdateContentStorageIdAfterSwitchingDriver, shouldWriteToContentAndReadItBack } from './testUtils';
